requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
absl-py=1.4.0=pypi_0
blas=1.0=openblas
bzip2=1.0.8=h620ffc9_4
ca-certificates=2023.01.10=hca03da5_0
cachetools=5.3.0=pypi_0
certifi=2023.5.7=pypi_0
cffi=1.15.1=py310h80987f9_3
charset-normalizer=3.1.0=pypi_0
fsspec=2023.5.0=pypi_0
future=0.18.3=py310hca03da5_0
google-auth=2.18.1=pypi_0
google-auth-oauthlib=1.0.0=pypi_0
grpcio=1.54.2=pypi_0
idna=3.4=pypi_0
libcxx=14.0.6=h848a8c0_0
libffi=3.4.4=hca03da5_0
libgfortran=5.0.0=11_3_0_hca03da5_28
libgfortran5=11.3.0=h009349e_28
libopenblas=0.3.21=h269037a_0
libprotobuf=3.20.3=h514c7bf_0
libuv=1.44.2=h80987f9_0
llvm-openmp=14.0.6=hc6e5704_0
markdown=3.4.3=pypi_0
markupsafe=2.1.2=pypi_0
mypy-extensions=1.0.0=pypi_0
ncurses=6.4=h313beb8_0
ninja=1.10.2=hca03da5_5
ninja-base=1.10.2=h525c30c_5
numpy=1.24.3=py310hb93e574_0
numpy-base=1.24.3=py310haf87e8b_0
oauthlib=3.2.2=pypi_0
openssl=1.1.1t=h1a28f6b_0
packaging=23.1=pypi_0
pip=23.0.1=py310hca03da5_0
protobuf=4.23.1=pypi_0
psutil=5.9.5=pypi_0
pyasn1=0.5.0=pypi_0
pyasn1-modules=0.3.0=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pyre-extensions=0.0.30=pypi_0
python=3.10.11=hc0d8a6c_2
pytorch=1.13.1=cpu_py310h9a5eeca_0
pyyaml=6.0=py310h80987f9_1
readline=8.2=h1a28f6b_0
requests=2.30.0=pypi_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.9=pypi_0
setuptools=66.0.0=py310hca03da5_0
six=1.16.0=pypi_0
sleef=3.5.1=hf27765b_1
sqlite=3.41.2=h80987f9_0
tensorboard=2.13.0=pypi_0
tensorboard-data-server=0.7.0=pypi_0
tk=8.6.12=hb8d0fd4_0
torch-summary=1.4.5=pypi_0
torcheval=0.0.6=pypi_0
torchtnt=0.1.0=pypi_0
tqdm=4.65.0=py310h33ce5c2_0
typing-extensions=4.5.0=py310hca03da5_0
typing-inspect=0.8.0=pypi_0
typing_extensions=4.5.0=py310hca03da5_0
tzdata=2023c=h04d1e81_0
urllib3=1.26.15=pypi_0
werkzeug=2.3.4=pypi_0
wheel=0.38.4=py310hca03da5_0
xz=5.4.2=h80987f9_0
yaml=0.2.5=h1a28f6b_0
zlib=1.2.13=h5a0b063_0
